+++ modules/graphical.nix
@@ -10,6 +10,13 @@
 
     services.xserver.enable = true;
     services.xserver.displayManager.gdm.enable = true;
+
+    sound.enable = true;
+    hardware.pulseaudio.enable = false;
+    services.pipewire.enable = true;
+    services.pipewire.alsa.enable = true;
+    services.pipewire.alsa.support32Bit = true;
+    services.pipewire.pulse.enable = true;
   };
 
   hmModule = { pkgs, ... }: {